% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potencia.R
\name{potencia2}
\alias{potencia2}
\title{Eleva ao quadrado numero}
\usage{
potencia2(x)
}
\arguments{
\item{x}{Um valor numerico e elevado ao quadrado}
}
\value{
O quadrado da entrada de dados
}
\description{
Eleva ao quadrado numero
}
